# [INF-87] faster `Building web server` step for discovery provider