--- a/creator-node/Dockerfile
+++ b/creator-node/Dockerfile
@@ -22,7 +22,6 @@ RUN echo 'http://dl-cdn.alpinelinux.org/alpine/v3.10/main' >> /etc/apk/repositor
     apk add postgresql=11.12-r0 && \
     apk add postgresql-client=11.12-r0 && \
     apk add postgresql-contrib=11.12-r0 && \
-    apk add postgresql-dev=11.12-r0 && \
     apk add postgresql-libs=11.12-r0 && \
     apk add python3 && \
     apk add python3-dev && \