--- a/discovery-provider/.dockerignore
+++ b/discovery-provider/.dockerignore
@@ -1,3 +1,4 @@
 venv
 logs
 *.swp
+node_modules