--- a/discovery-provider/scripts/openresty_log_convertor.py
+++ b/discovery-provider/scripts/openresty_log_convertor.py
@@ -1,5 +1,5 @@
-import sys
 import json
+import sys
 
 for line in sys.stdin:
     message = line.strip()