… discovery-provider/scripts/init-db.sh
… discovery-provider/scripts/search_quality.py
… discovery-provider/scripts/seed.py
… discovery-provider/scripts/wait